Eliminate phasing
In phase 1 C3 will already need to support all states. The phase 2 states can only be optional for C2, because (1) the schematron allows all these states to be sent and we are not creating a new document type id with new schematron rules. And (2) it makes little sense for C2 to implement only a subset of states. What benefit would that have, doing the same project twice?

That is different for C1/C4. C1 can choose whether or not to advertise the Invoice Response document and C4 can then either send it or not. Advertising Invoice Response for C1 is optional. We are (October 2021) years away from being able to force all senders to process Invoice Response documents. But if C1 chooses to advertise, they must consume all states.

So I propose we eliminate the phasing. The project with Gerard Bottemanne already showed this is possible.