MLR/AP not mandatory
I hope the authors realize that by making the MLR/AP mandatory, we are doubling the number of Peppol messages overnight. Not something to take lightly.

The MLR/RE (not AP) is an important message and that is missing from the Dutch ecosystem. Having a way to reject invoices other than through the AccountingSupplierParty contact is important. So MLR/RE solves a big problem. Should be mandatory for all SPs, both sending and receiving.

But regarding MLR/AP I have several questions/points that I think need to be answered/investigated before we can decide to proceed with making this mandatory.

1. What problem does MLR/AP solve?
I can't see any, please enlighten me.

2. What is the legal status of the AS4 signature in relation to MLR/AP?
In The Netherlands we use mainly the "ontvangsttheorie" ("receive theory") which means for a document to have legal effect ("rechtskracht") it must have been received. But that term "received" is objectified. If a letter is sent to you home address but you are on 6-month world tour, it is considered received. Even if you have not read it. If you own a Postbox and have your mail sent there, but you never empty it, it is considered received. If you have a Postbox and a service that forwards all this mail to your office address, but this service fails to do so, the document is considered received. Because you are responsible for that service from the Postbox to your office.

With Peppol it is much the same. When C2 sends to C3, the document has been received. And C2 has a digital signature to prove it. Like a registered letter. And if C3 fails to send the document to C4, that is the responsibility of C4, because C4 engages C3's services.

So far it is clear. But with the introduction of a mandatory MLR/AP, are we saying that we are changing this model? No longer using wat is current practice? Before muddying the waters by making MLR/AP mandatory I recommend the NPa consult with legal counsel to obtain an impact assessment. If we are solving a problem, are we not creating another?

3. Volume
Since we are doubling the volume overnight, I recommend that the NPa consult with the SPs to make an impact assessment. We are not only talking about naked volume of (yes, small) documents, but also the support and rework required in case of delivery problems. Every document adds a little bit to that cost. Probably an investment SPs are willing to make, if with MLR/AP we are solving a real problem.